Sets up enviroment and beginning of better api system
I am having problems now with the promises not being
fulfilled from the api request. That leads to me not
being able to call on the arrays when I need them in my
basic component. I guess this is why rengorum has all
of the '...COMPLETED' flags. I will look into that next.