Adds redux async call to the backend
It is still really simple but I set up a store and
all of the other mechanisms for redux to get a list
of the users (i.e. redditors) from the api and list
them on the 'users' page of the front end.

Next, I think I need to seriously think about what the
overall design of the site should be like. That is, I need
to go look at reddit and get a basic homepage idea going. I
will implement that next.