Performs api call for generating post list
This is the first completed example I have done. On the homepage
there is a list of all posts from the backend. They are styled,
very basically, in semantic ui react.

The next step is figuring out how to pare down the list of posts
to get only relevant, i.e. 'popular' or 'new'. Perhaps that is best
done on the backend. I will take a look at the reddit api and see
if they provide something like that or if they do it on the frontend.