Moved modal component to bootstrap from semantic-ui
I have found that having both libraries is becoming a
problem with css. I tried first to go with semantic since
there is so much more styling available but I had a real
headache making the navbar workout and I think bootstrap
is just a little simpler. The point here is to illustrate
that I can work with react and django, not that I am a
great designer.

Going forward I will be uninstalling semantic-ui.