Enables login with user input
This was a major reworking of the Auth and Auth Modal stuff.
I split the login modal into a general modal component and a
login form. The main modal container now shows a particualar
and more specific modal container, e.g. LoginModalContainer,
which then pairs the modal and login form pairs.

I would next like to take care of the asynchronicity that occurs
during login and put a spinner on the modal when we are requesting
the token. The we should probably implement logout functionality too.