Started on user reg and then wrote api middleware
Pretty poor source control management this is two large
commits really. I started to wire everything up for
user registration and it was such a pain in the butt that
I decided it was a good time to write some middleware
to minimize the amount of similar code that needed to be
copied.

At this point I have named the middleware testMiddleWare. And it
really was a test. I created a test action that is dispatched when
the user registration button is clicked and is hard-coded to login
a test profile.

The next step is to rewrite the actual user login functionality to use
the middleware.