Creates apiMiddleware and wires up user login
One outstanding issue is closing the login modal once the
user is logged in. That was handled directly in the old async
action creator before. I was thinking that I could add a
'on success action' funcitonality to the apiMiddleware. Alternatively
it may be best to return a promise from the middleware and just tack on a
.then(close the modal) wherever the new action creator is dispatched.