Adds error message for login
I ended up adding an onFailure action creator in the middleware
to make this happen. So when you fail to login you fire an
action that makes the UserAuth state have an error message that
relays the error from the response.

This might work but I'm leaning more toward just
having the USER_AUTH_FAILURE error make the
modal state have an error message. Its a pain to pass
a success message and a error message (stupidly named warningMessage)
through a few components. The next step is making that a little
more managable and then we will see how it looks.