Adds orderBy funcitonality to reddit posts api requests
At this point there are two different api request functions
and corresponding action callAPI functions.
One for a specific subreddit and one for posts from all
subreddits. I am thinking that it may be best to make the
current sub a state (redux) variable when we are looking at
a post list. Then set it to null when we want all of the posts.

The thing I am not sure about it how that will all play with
router. For now I want to move forward and will leave them separate
but I will look into how to route this next.