Adds new and popular links to navbar dropdown
Currently the change the url but that doesn't re-render the
PostContainerList. The issue, i think, is that there is no
real state in that component to change. I saw a suggestion
to use an HOC, beyond the container itself, to fix this issue.
I am still trying to figure out how that would work.