Makes postList branch in store a child of subreddit
subreddit is now a child of the main root-reducer. Since
the postlists are only loaded in conjunction with a subreddit
being displayed it makes more sense to make postList a sub-object
of subreddit.

We will still make two separate api calls, one for the sub
details (without links to the sub posts) and one specifically
to get the post of the sub. They are separated so that later
we can implement some of the reddit filtering (e.g. new, hot,
popular) while on the same subreddit page and just fire off
a new api request for the posts. This has the advantage of
modularity but more importantly we can let the backend do the
sorting for us, potentially with some async stuff if
necessary, (though I envision that async stuff
more for the subreddits 'new' and 'popular').