Adds a Subreddit container and component
Previuosly we just had a post list doing all the work.
Now the PostList component (And container) is a child of
the Subreddit container. This makes the code a lot easier to
follow and allows the names of the components to make more sense.
Furthermore it makes it easy to seperate the api requests for
sub details and post lists (e.g. we can now do the new/popular
post sorting without reloading the entire subreddit view.

The next step is figuring out which of the sub detail data we
need in either the post list or the sidebar and loading that
with the Sub detail api request action.