Fixes issue with home subreddit name
There is a problem with the home subreddit. It will be unique
to each user and I want it to be at the root url. Currently I have
the action creator for the subreddit detail request setup to fire
a synchronous action that just sets the subreddit title as 'home'.

The post list fetching is still wired to return the entire list of
all of the subs for the home page. That might work okay for
an anonymous user but for someone who is authenticated we want only
the subs to which they are subsribed to populate the list.