Adds subreddit subscribe functionality
Now the button on the subreddit page is hooked up to the
api request for subscription. It works to subscribe.
There are two problems.

First, although we have a
'subscribed' state variable that works I havent done
anything with it yet (i.e. change subscribe button to unsubscribe).
Second when a user subscribes we need to fire off another
user details request to update the list of subreddit subscriptions.