Refactors design of redux state to remove superfluous 'subscribed'
Instead of having a subreddit.subscribed variable now we just
pass userAuth.subs (titles map) to the subreddit container.
The subreddit components can use the current subreddit title and
this list of the current users subscriptions to figure out what
we need.

I removed the subreddit.subscribed varibale from the sr reducer and
got rid of the subscriptionCheck action. The next main hurdle is to
be sure that the userAuth sub list still gets updated when we have
successfully subscribed.