Improves handling of subscription tracking
This build on the redux reformulation. Now when a user subscribes
we make the subscription request (either a subscription or
unsubscription) and then, upon the success of that request
we make a request for the updated user details. The bulk
of this commit is the boilerplate for that latter api request.
I think this is a better way to go than keeping track in just the
front end by updating the redux store and hoping that it
agrees with the backend.

Another important part of this commit is changes to the apiMiddleware.
The 'types.success' can now be a thunk. Instead of just firing off
an action with the type of types.success now we evalute the thunk
with the data, getState and dispatch passed to it. Its alot like
redux thunk. This is where we can dispatch both the update to the
subreddit.subscriptionLoading action and the api request action
for the updated user details.
In the future I think I would like to move all of the 'types'
fields to this structure.