A first step in generalizing the subreddit components for createPost
I noticed that in the design of reddit they usually have only a
single flex box with two columns for all content, regardless
of what exaclty is being displayed. I already noted and implemented
that for subreddit post lists. Now I would like to repurpose
the subredditContainer for creating posts within a subreddit. I think
it can remain named subredditContainer because whether it holds
postlists or the createpost form it still will pertain to a particular
subreddit.

This commit takes a first step towards that in permitting
a styling change at the top level for the two flex columns. It
depends currently on whether there is a '/createPost' appended
to the subreddit URL. This switch will be used to make the
subreddit container show a createpost form as the primary content
(as opposed to a postlist) . Fortunately the sidebar is
a separate concern but I think it can remain pretty much the same for
both use cases.