Conditionally render createpost container in subreddit primary content
I had some trouble with the HOC leading to an excessive number
of updates, for some reason the post list is continually remounted
and the api call in componentDidMount causes an infinite loop
of changing props. For now I am just using an if statement instead
but it's a good learning moment so leaving the commented code in
there to figure it out later.