Refactors root design
In order to continue reusing the basic flex-box primary and sidebar
structure the root has been refactored to call Routes with
the render= functionality (as opposed to component=). This
makes it possible to pass different props for each route, even when we
use the same component, e.g. SubredditContainer. These
props are needed to indicate some styling choices and which
component should be loading in the primary slot, e.g. CreatePost
vs PostList.

Soon we are adding PostDetail to the list as well.