Moves parent fullname creation to api function
Previuosly I was forming the parent full name (i.e. t1_<pk>)
in the component. I think it is better to locate that
in the api funciton and just pass the parent pk and
a bool to indicate whether the parent is another comment (t1_)
or a post (t2_). It's really the business of the api interface
not the component.