Creates comment voting
Create the api interface stuff and hooks into the comment
components. I think that going forward I should make a
separate component with its own container for a voter.
It can be used for both the
comments and the posts. Currently I had to prop drill 4 levels
with the dispatch function for the vote api action creator.