Adds Voter component and container
This links directly to a vote api action creator. It takes
the initialDisplayState as a parameter. That is passed from the
comment which is provided with information about whether the
currently authenticated user has voted on that comment yet.

The voter is wired up to use the displayState stuff and change
appropriately. I need now to refactor the comment tree request to
include a username (if a user is authenticated) to the url of
the GET request.