Updates the comment creation process
This normalized structure makes it easier to do the reddit
trick of bringin a nascent comment to the top of it's siblings.

It was a little trickier to deal with in the reducer than I
initially planned but it is still simpler than my original
attempt that passed the nascent comment through the components.