Simplifies handling of vote state
Previously we had the vote state being managed by the state
of the voter component and the comment info in the redux store.
Obviously not a great idea, this had the advantage of making
the color change upon voting really quick. When it's managed
by the redux store value it is slower because that update waits
untill it hears a success response from the backend.

If that is really important then I will just add a new
action e.g. superficial_comment_vote that doesn't wait for a
request and just changes the redux state imediatly upon a vote.