Bug fix for new comment creation
When a new comment is created we take the success response from
the backend and set that into state.comments.commentsById.
The issue was that in the response the poster is refered to
by username, not pk. When we retrieve the trees they undergo
normalization which takes a nested user and leaves only a pk.
This mismatch has us looking up an id of 'Bobby'.

For the fix I just grabbed the userAuth pk and set that during
the addition of the new comment (i.e. when we add it to the top
of the lists, before the list is fetched again and it goes
to it's place.