Adds upvotes prop to the voter
In reddit they have the number of votes for a post inbetween the
arrows in the voter. They just have adjacent arrows in the comment
list though. This upvotes prop is optional, so when the voter is
created for the post segments we can just pass them there
and ignore them in the comment segments.