Adds comment body component
This is a sub-component of Comment. I noticed that for both
a delted comment and a comment being edited we needed to
replace/remove the comment body and buttons line. I wrapped them
into a component to facilitate that.

Additionally I added the prop-types library and put prop types
in for the new component. I would like to slowly start adding them
to all of the components.