Fixes bug with wrong comments loading on post detail
The comment list api request action would be dispatched first
since it is the most deeply nested component and we
call the api actions from `componentDidMount`. It would
use the wrong postPk, whichever post we happened to be on
last. Once
the PostDetail component calls `componentDidMount` and
resets the postPk within the redux store the CommentList was
remounted (because we had a Loader mounted in the interim).
Then the CommentList.componentDidMount calls the api action
again with the proper postPk. Sometimes with
particular posts this initial request would be fulfilled after
the second one, i.e. the `FETCH_..._SUCCESS` action. So we would
be stuck with the previous post's comments in our new PostDetail
view.

To fix this I used withRouter and only call the fetchCommentList
api when the router and redux agree on the post pk within the
CommentList.componentDidMount function. That avoids the original call
because at that point the router has the correct post id and
the redux store still has the old one.

Another option would be to somehow fire an action immediately upon
the router changing. Or maybe it would be best to not even store
the postPk in the redux store, we kind of run into this 'two sources
of truth' problem. The problem with that is that I don't
want to have to pass the router props (or use withRouter) all over
the place, the postPk is a popular prop and it's nice to be able to get
it from redux. For now I will just be cognizant of the issue.