Refactors componentDidMount on CommentList again
This is related to the same issue that I wrote the blog
post about.
Now that I am handling the loading of the CommentList within
the component itself I would either need to switch back
to componentDidUpdate or just use the react-router params
to find it right the first time in DidMount. I went with the
second.