Adds subreddit panel and list in search results
I decided to do it right and make a SubredditPanelContainer
then we can handle the subreddits just like the posts. Pass an
allSubreddits list of ids to the search results and then
create a SubredditPanel for each item in the list. They are
a lot simpler than the PostPanels though.