Simplifies the comment editor container logic
I had previously concocted this mess using mergeprops so that
when used to create a new comment on a post the CommentEditorContainer
could operate in isolation and grab the postPk using mapStateToProps.

I have since decided that it's pretty much always a better bet to just
pass the postPk down through a few components. When I looked into it
the postPk was already available in the CommentTreeList which
calls CommentEditorContainer, so it was easier than I thought.