Bug fix for create functionality in CommentEditorContainer
I messed up with my recent clean up effort. I pass the pk for the case
when we are creating the comment but I forgot to pass the same prop
name from both a post reply creation and a comment reply creation.
Now either way we just pass a `parentPk` prop and along with the
`rootComment` prop enough information is present to determine
the fn of the parent to pass to the backend.