Install after node update
This updates quite a few packages. Most importantly this
fixes the issue that I was having with run build. I was
getting a 'cannot read search property of undefined' errror.

I thought it may have been a bug in my code since there is
a slice of redux state called search. I'm not sure about that anymore.
Whatever the case it builds on my dev machine now. I am going to
upload this and clone it to the ec2.