Updates design of modal container to avoid rerendering/mounting
Previously the Modal component itself was at the leaf components.
One for login and one for register. I moved that to the very top
of the general (i.e. for both login and register) UserAuthModal
container. This allows the Modal component to remain mounted
when switching between login and register through the
'Already a ...' links on the bottom of the forms.