Creates a custom navbar
The impetus for moving away from bootstrap it to make it a little
more mobile freindly. It was getting too hard to customize anything.
It has been completely remade by hand, albeit with bootstrap
dropdowns. All that remains to inplement is the hamburgering
of the search and user dropdowns. I think that can be handled
in the subcomponents themselves.