Forces updates on postlist and comment tree with login/out
It is necessary to reload on user auth because there is user
specific information present in the lists of posts and comments,
e.g. if they voted on something the voter will be red or blue.

Previously we just redirected to the home screen on login
(forcing an update unless the user was already there) and did
nothing on logout. This doesn't work very nicely so now the
PostListContainer and CommentTreeContainer check to see if the
authed user has been updated in the redux store and will
update their lists if need be.