Fully moved post list handling to the postlist slice of reducer
Previously we had allPosts for the user profile in both
the userProfile and postList slice of the redux store. Besides
being a bad multi source of truth design it led to the
user profile container pulling allPosts from the userProfile
slice while the postList within the profile view was pulling
from the postList slice. This led to mismatches, e.g. when
you search and then use the browser back button, between posts
in the two slices that resulted in 'undefined does not have property
...' type errors.