Hides the user page in search result
It's an odd one to show. There is really nothing to display other
than the username, maybe the cakeday/karma. I could not find what
one looks like on the real reddit search results page to model it
after either.

For now I will just hide it. Maybe later if we add an avatar picture
it will be more realistic to add the results list back in.