Handles normalizing the nested posts from a list of user profile comments
We need to have post information for the comments listed in a
user profile page. They are used for mentioning which post it's
from and so that the post can be linked. The backend was updated
to include this information as a nested object in the comment
json.

This commit takes that nested post object, for each comment, and
normlizes it for placement into the standard redux design with
commentPostsById.