Adds api and components for a next age loader
This is a special panel loader that will sit at the bottom of the
post list. When the loader enters the viewport it triggers a request
for the next page of posts.

This commit adds the request action/api and the container. The
component itself is currently just a loader. I have yet to implement
the 'on entry to viewport trigger the fetch action' funcitonality.