Completes the scrolling load functionality for post list
I hooked up the api with the NextPageLoader using a
3rd party package. It loads new posts as expected and adds
them to the bottom of the current post list without remounting
the post list itself.

It also avoids showing the NextPageLoader on lists where there
is nothing more to show. That is accomplished by tying the loader to
the existence of a nextPageUrl in the redux state.