Bug fix for loading the wrong posts in user profile view
When I implemented the nextPageLoader I forgot to make special
arrangements for the postList that is used in the user profile view.

It was just loading new posts that had nothing to do with the
user profile in question automatically. I put an optional
prop to avoid rendering the nextPageLoader (and thus making
new post requests) into postList, it defaults to true.