Links to post detail from comment post line in user profile
I would like to do something similar for the comments but I don't
have a view set up to handle when we load a post detail for a
particular comment. Reddit does it by only loading the parents.

I think that would be a fun challenge but not right now.